more uint32 fixes

+++ api/oc_knx_fp.h
@@ -251,9 +251,9 @@ void oc_register_group_multicasts();
  *
  * @param group_address The group_address from the group object table
  * @return the grpid matching the group_address the table publisher table
- *  or -1 if not found
+ *  or 0 if not found
  */
-int oc_find_grpid_in_publisher_table(int group_address);
+uint32_t oc_find_grpid_in_publisher_table(uint32_t group_address);
 
 /**
  * @brief find the grpid from the group_address in the recipient table
@@ -262,9 +262,9 @@ int oc_find_grpid_in_publisher_table(int group_address);
  *
  * @param group_address The group_address from the group object table
  * @return the grpid matching the group_address the table publisher table
- *  or -1 if not found
+ *  or 0 if not found
  */
-int oc_find_grpid_in_recipient_table(int group_address);
+uint32_t oc_find_grpid_in_recipient_table(uint32_t group_address);
 
 /**
  * @brief initializes the data points at initialization
@@ -287,7 +287,7 @@ int oc_core_find_index_in_group_object_table_from_id(int id);
  * @param group_address the group address
  * @return int the index in the table or -1
  */
-int oc_core_find_group_object_table_index(int group_address);
+int oc_core_find_group_object_table_index(uint32_t group_address);
 
 /**
  * @brief find next index in the group address table
@@ -296,7 +296,7 @@ int oc_core_find_group_object_table_index(int group_address);
  * @param cur_index  the current index to start from.
  * @return int the index in the table or -1
  */
-int oc_core_find_next_group_object_table_index(int group_address,
+int oc_core_find_next_group_object_table_index(uint32_t group_address,
                                                int cur_index);
 
 /**
@@ -406,7 +406,7 @@ void oc_delete_group_rp_table();
  * @return false is not part of the recipient entry
  */
 bool oc_core_check_recipient_index_on_group_address(int index,
-                                                    int group_address);
+                                                    uint32_t group_address);
 
 /**
  * @brief get the destination (path or url) of the recipient table at index
@@ -575,8 +575,9 @@ void oc_free_knx_fp_resources(size_t device_index);
  * @param scope the address scope
  * @return oc_endpoint_t the modified endpoint
  */
-oc_endpoint_t oc_create_multicast_group_address(oc_endpoint_t in, int group_nr,
-                                                int iid, int scope);
+oc_endpoint_t oc_create_multicast_group_address(oc_endpoint_t in,
+                                                uint32_t group_nr,
+                                                int64_t iid, int scope);
 
 /**
  * @brief subscribe to a multicast address, defined by group number and
@@ -588,7 +589,7 @@ oc_endpoint_t oc_create_multicast_group_address(oc_endpoint_t in, int group_nr,
  * @param iid the installation id
  * @param scope the address scope
  */
-void subscribe_group_to_multicast(int group_nr, int iid, int scope);
+void subscribe_group_to_multicast(uint32_t group_nr, int64_t iid, int scope);
 
 /**
  * @brief unsubscribe to a multicast address, defined by group number and
@@ -600,7 +601,7 @@ void subscribe_group_to_multicast(int group_nr, int iid, int scope);
  * @param iid the installation id
  * @param scope the address scope
  */
-void unsubscribe_group_to_multicast(int group_nr, int iid, int scope);
+void unsubscribe_group_to_multicast(uint32_t group_nr, int64_t iid, int scope);
 
 #ifdef __cplusplus
 }
